requirements.txt
absl-py==1.1.0
actionlib==1.13.2
angles==1.9.13
beautifulsoup4==4.11.1
bondpy==1.8.6
boto==2.49.0
cachetools==5.2.0
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2022.5.18.1
charset-normalizer==2.0.12
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cv-bridge==1.16.0
cycler==0.11.0
Cython==0.29.30
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.2
easydict==1.9
filelock==3.7.1
flake8==4.0.1
fonttools==4.33.3
future==0.18.2
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gdown==4.4.0
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
h5py==3.7.0
idna==3.3
image-geometry==1.16.0
imageio==2.19.3
importlib-metadata==4.11.4
interactive-markers==1.12.0
isort==4.3.21
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.2
laser_geometry==1.6.7
Markdown==3.3.7
matplotlib==3.5.2
mccabe==0.6.1
message-filters==1.15.14
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
openpifpaf==0.13.4
packaging==21.3
pandas==1.4.2
-e git+ssh://git@github.com/theoh-io/Perception-Pipeline.git@4d287a90e580a12fefb0b918e1139448350cf0fd#egg=perceptionloomo&subdirectory=python
Pillow==9.1.1
pkg_resources==0.0.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
PySocks==1.7.1
pysparkling==0.6.1
python-dateutil==2.8.2
python-json-logger==2.0.2
python-qt-binding==0.4.4
pytz==2022.1
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.27.1
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.2
rqt_gui_py==0.5.2
rqt_image_view==0.4.16
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.2
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.2
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.8
rviz==1.14.14
scipy==1.8.1
seaborn==0.11.2
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
soupsieve==2.3.2.post1
tb-nightly==2.10.0a20220603
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
topic-tools==1.15.14
torch==1.11.0
## !! Could not determine repository location
-e /home/theo/Perception-Pipeline/python/src/perceptionloomo/libs/deep-person-reid
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
xacro==1.14.13
yacs==0.1.8
yapf==0.32.0
zipp==3.8.0
